fix the file name to meet pylint rule
