+++ test/strategy/test_bayesian_1x.py
