+++ test/strategy/test_mse_v2_2x.py
