+++ test/strategy/test_sigopt_1x.py
